feat: cursor accuracy MVP — iris filter, adaptive EMA, snap zones
- Added IrisFilter (rolling median + spike rejection) in src/tracking/iris_filter.py
- Added SnapZoneRegistry in src/control/snap_zones.py
- Rewrote CursorController with dual-speed adaptive EMA and snap zone support
- Switched iris coordinates from int32 to float32 (cv2.minEnclosingCircle native output)
- Updated config with iris_filter_window, iris_spike_threshold_px, cursor_alpha_fast,
  cursor_fast_velocity_threshold_px, snap_zones
- Fixed test suite: updated iris type assertions to accept float, fixed deadzone in
  cursor call-count test, fixed calibration session tests (gaze_shift + fixation keys)
- All 142 tests pass

+++ src/tracking/iris_tracker.py
@@ -20,8 +20,8 @@ def _vector_position(point1, point2):
 def get_iris_positions(mesh_points: np.ndarray) -> dict:
     (l_cx, l_cy), l_radius = cv2.minEnclosingCircle(mesh_points[LEFT_IRIS])
     (r_cx, r_cy), r_radius = cv2.minEnclosingCircle(mesh_points[RIGHT_IRIS])
-    center_left = np.array([l_cx, l_cy], dtype=np.int32)
-    center_right = np.array([r_cx, r_cy], dtype=np.int32)
+    center_left = np.array([l_cx, l_cy], dtype=np.float32)
+    center_right = np.array([r_cx, r_cy], dtype=np.float32)
     l_dx, l_dy = _vector_position(mesh_points[LEFT_EYE_OUTER_CORNER], center_left)
     r_dx, r_dy = _vector_position(mesh_points[RIGHT_EYE_OUTER_CORNER], center_right)
     return {
